deprecate implied recursion
Up until now, recursing into a subdirectory could be specified by just
putting the name of the directory in .dfminstall, like this:

-----------------
.ssh                 <--- causes dfm to recurse into the .ssh dir
README.md skip
t skip
-----------------

This is slighly confusing, because to just skip a file, a 'skip' keyword
was required.  This change makes recursion behave the same as skipping.
The above file should now look like this:

-----------------
.ssh recurse         <--- causes dfm to recurse into the .ssh dir
README.md skip
t skip
-----------------

A deprecation warning will be displayed if implied recursion is used.