Replaced relative paths with absolute paths in the install_files recursion code to fix mkdir mishaps
The tests for links and directories in install_files recursion code were relative paths so
when more than one directory was specified for recursion in a .dfminstall file the tests
would be run for the wrong locations.

This fix uses absolute paths which ensures the locations are correct.

This fix also adds the absolute paths to debug output for the recursion code in both
install_files and uninstall_files for clarity when debugging.